requirements.txt
Flask==2.2.5
Werkzeug==2.2.3
beautifulsoup4==4.13.4
pandas==2.3.1
Pillow==11.3.0
pytesseract==0.3.13
python-dotenv==1.1.1
requests==2.32.4
tqdm==4.67.1
urllib3==1.26.11
gunicorn==20.1.0
wcwidth==0.2.13
webdriver-manager==4.0.2
websocket-client==1.8.0
wsproto==1.2.0
xgboost==3.0.0
pyodbc
